requirements.txt
openai
python-dotenv
wikipedia
tiktoken
neo4j
streamlit
# sentence_transformers==2.2.2
Pillow
fastapi
PyPDF2
torch==2.0.1
pydantic
uvicorn
sse-starlette
boto3
# missing from the langchain base image?
langchain-openai
langchain-community
langchain-google-genai
langchain-huggingface
# langchain_huggingface
javalang==0.13.0
unidiff~=0.7.4
langchain-ollama
